backend/requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.6.0
boto3==1.20.7
botocore==1.23.7
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coolsms-python-sdk==2.0.3
cryptography==36.0.1
daphne==3.0.2
distlib==0.3.5
Django==4.0.6
django-cors-headers==3.10.0
django-filter==21.1
django-hosts==5.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.0
filelock==3.7.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
identify==2.5.2
idna==3.3
importlib-resources==5.4.0
incremental==21.3.0
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.3
msgpack==1.0.3
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.2
prompt-toolkit==3.0.28
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
s3transfer==0.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
Twisted==22.1.0
txaio==21.2.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.16.1
wcwidth==0.2.5
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
pycryptodome==3.15.0